AuraLinkBackend/requirements.txt
flask
paho-mqtt
langchain
langchain-openai
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv

